requirements.txt
pydantic
pyyaml
requests
bentoml
tqdm
openai
numpy
Pillow
aesthetic_predictor_v2_5
